Async Log Event Support
It is possible that customers are writing logs asycnrhonously in their
applications. If new relic is capturing those logs, we need to ensure
that the state of the application is not compromised, and that the
customer's expectaions for which logs are being captured are consistent
with what the application is actually doing.